// apps/trade-pro/src/pages/StoreManagement/StockTransferNotes/types.ts
export type TLoadORderHistory = {
  Id: string;
  RequisitionDetailId: string;
  OrderDate: Date;
  OrderNo: number;
  SourceId: number;
  SourceLocation: string;
  DestinationId: number;
  DestinationLocation: string;
  ItemId: number;
  ItemName: string;
  ItemUomId: number;
  PackUom: string;
  PackEquivalent: string | any;
  ReqQty: number;
  ApprovedQty: number;
  ReqRate: number;
  ReqWeight: number;
  StockWeight: number;
  ReqAmount: number;
  RemarksHeader: string;
  WareHouseName: string;
  AvailableStock: number;
  AvailableWeight: number;
  BillWeight: number;
  IssuedAmount: number;
  IssuedQty: number;
  IssuedRate: number;
  ExpenseAmount: number;
  ItemNetAmount: number;
  ActionTypeId: number;
  WarehouseId: number;
  LineId: number;
  WsRmRequisitionPoDetailId: number;
  WsRmRequisitionPoId: number;
  WsRmStockTransferNotesId: number;
  RemarksDetail: string;
  UOMCode: string;
  RequisitionNo: number;
};
// ...
export type TWsRmStockTransferNotesExpensesList = {
  ItemId: number;
  ItemName: string;
  Qty: number;
  Rate: number;
  Amount: number;
  Remarks: string;
};